Remove the second instance of the app and rename its files.
