+++ app/main.py
@@ -1,52 +1,74 @@
 import subprocess
-import time
 from app.telegram import send_telegram_message
 from app.viber import send_viber_message
-from app.vars import plants
+from app.vars import Vars, allvars
 
 
-def ping_server(ip_addr):
-    command = ["ping", "-n", "2", ip_addr,]
-    subprocess.run(
-        ["chcp", "437"],
-        shell=True,
-        stdout=subprocess.DEVNULL,
-    )
-    try:
-        output = subprocess.check_output(
-            command,
-            stderr=subprocess.STDOUT,
-            encoding='cp866',
-            creationflags=subprocess.CREATE_NO_WINDOW
-        )
-    except subprocess.CalledProcessError:
-        return True
+class AuditShields:
 
-    if "TTL" in output:
-        return False
-    elif "100%" in output:
-        return True
+    def __init__(self, config_vars: Vars) -> None:
+        self.vars = config_vars
+        self.shields_out = {}
+        self.messages_sent = {}
+
+    @staticmethod
+    def ping_host(ip_address: str) -> bool:
+        """ Checks a single host if it's up or out """
+        command = ["ping", "-n", "1", ip_address, ]
+        subprocess.run(["chcp", "437"], shell=True, stdout=subprocess.DEVNULL, )
+        try:
+            output = subprocess.check_output(
+                command,
+                stderr=subprocess.STDOUT,
+                encoding='cp866',
+                creationflags=subprocess.CREATE_NO_WINDOW
+            )
+        except subprocess.CalledProcessError as err:
+            output = err.output
+
+        if "TTL" in output:
+            return False
+        elif "100%" in output:
+            return True
+
+    def is_network_out(self, network: str) -> bool:
+        """ Checks an always working host to make sure its network works """
+        checking_host_ip = self.vars.hosts[f"{network} SOURCE"]["in_touch"]
+        is_out = self.ping_host(checking_host_ip)
+        return is_out
+
+    def check_shields(self, network: str) -> dict:
+        for shield, hosts in self.vars.hosts.items():
+            if network in shield and "SOURCE" not in shield:
+                hosts_out = [self.ping_host(host) for host in hosts.values()]
+                self.shields_out.update({shield: all(hosts_out)})
+        return self.shields_out
+
+    def form_alarm_message(self) -> str:
+        """ Composes an alarm message text regarding the certain equipment alarm """
+        message_text = ""
+        for shield, status in self.shields_out.items():
+            if not status:
+                continue
+            if not self.vars.sendings[shield]:
+                self.vars.sendings[shield] = True
+                message_text += f"{self.vars.messages[shield]} \n"
+        return message_text
+
+    @staticmethod
+    def send_alarm_message(text: str) -> None:
+        """ Sends the alarm message to proper Telegram and Viber users """
+        if text:
+            text = f"Alarm!\n{text}"
+            send_telegram_message(text)
+            send_viber_message(text)
 
 
-def is_server_out(vent_units):
-    for shield, hosts in vent_units.hosts.items():
-
-        servers_out = [ping_server(host) for host in hosts.values()]
-        servers_out += [ping_server(host) for host in hosts.values()]
-
-        if None in servers_out:
-            continue
-
-        elif all(servers_out) and not vent_units.sendings[shield]:
-            alarm_message_text = vent_units.messages[shield]
-            send_telegram_message(alarm_message_text)
-            send_viber_message(alarm_message_text)
-            vent_units.sendings[shield] = True
-
-        elif not all(servers_out):
-            vent_units.sendings[shield] = False
-
+auditor = AuditShields(allvars)
 
 while True:
-    is_server_out(plants)
-    time.sleep(30)
+    for net in ["WIFI", "DLAN", "INET"]:
+        if auditor.is_network_out(net):
+            auditor.check_shields(net)
+    message = auditor.form_alarm_message()
+    auditor.send_alarm_message(message)
