--- app_v2/telegram_.py
@@ -1,14 +1,0 @@
-import telebot
-from app_v2.vars_ import allvars
-
-telegramtoken = allvars.telegram_configs["TELEGRAMTOKEN"]
-bot = telebot.TeleBot(telegramtoken)
-
-def send_telegram_message(message_text):
-    for user in allvars.telegram_users.values():
-        bot.send_message(user, message_text)
-
-
-if __name__ == "__main__":
-    admin = allvars.telegram_users["admin"]
-    bot.send_message(admin, "The bot checking!")
